+++ app/controllers/forms/base_controller.rb
@@ -7,7 +7,7 @@ module Forms
       LogEventService.log_form_start(logging_context) unless mode.preview?
     end
 
-    rescue_from ActiveResource::ResourceNotFound, StepFactory::PageNotFoundError do
+    rescue_from ActiveResource::ResourceNotFound, Flow::StepFactory::PageNotFoundError do
       render template: "errors/not_found", status: :not_found
     end
