+++ config/routes.rb
@@ -23,7 +23,7 @@ Rails.application.routes.draw do
       get "/submitted" => "forms/submitted#submitted", as: :form_submitted
       get "/privacy" => "forms/privacy_page#show", as: :form_privacy
       get "/:page_slug/change" => "forms/page#show", as: :form_change_answer, defaults: { changing_existing_answer: true }
-      get "/:page_slug" => "forms/page#show", constraints: { page_slug: StepFactory::PAGE_SLUG_REGEX }, as: :form_page
+      get "/:page_slug" => "forms/page#show", constraints: { page_slug: Flow::StepFactory::PAGE_SLUG_REGEX }, as: :form_page
       post "/:page_slug" => "forms/page#save", as: :save_form_page
 
       get "/repeat-submission" => "forms/base#error_repeat_submission", as: :error_repeat_submission, via: :all
